tegra: ivc: update channel header format
Drop the cbuf structure entirely for use in IVC, as it is not
appropriate for the task. The main issues being that counters that are
expected to be updated simultaneously share cache lines and most other
fields are unusable because the peer could have tampered with them.

Switch to a simpler structure containing cache-aligned, free-running
counters that only reflect the number of frames read and written by
each end. The actual position of the next frame in shared memory is
maintained locally. This structure also contains a state word in the
sender area that will be used for a future reset protocol. As a result
of the new counter scheme, we no longer need to waste a frame to
implement the full condition.

The required alignment is currently 64 bytes, which is the cache line
size on the ARMv8 implementations we plan to support.

This is a change to the binary IVC protocol that must be synchronized
with all implementations.

Bug 1535037
Bug 1559020

Reviewed-on: http://git-master/r/562999
Tested-by: Bahadir Balban <bbalban@nvidia.com>

+++ include/linux/tegra-ivc.h
@@ -32,6 +32,8 @@
 #ifndef __TEGRA_IVC_H
 #define __TEGRA_IVC_H
 
+#include <linux/types.h>
+
 struct device_node;
 
 /* in kernel interfaces */
@@ -95,7 +97,7 @@ int tegra_hv_ivc_unreserve(struct tegra_hv_ivc_cookie *ivck);
  */
 int tegra_hv_ivc_write(struct tegra_hv_ivc_cookie *ivck, const void *buf,
 		int size);
-int tegra_ivc_write(struct ivc *ivc, const void *buf, int size);
+int tegra_ivc_write(struct ivc *ivc, const void *buf, size_t size);
 
 /**
  * ivc_hv_ivc_read - Reads a frame from the IVC queue
@@ -108,7 +110,7 @@ int tegra_ivc_write(struct ivc *ivc, const void *buf, int size);
  * Returns size on success and an error code otherwise
  */
 int tegra_hv_ivc_read(struct tegra_hv_ivc_cookie *ivck, void *buf, int size);
-int tegra_ivc_read(struct ivc *ivc, void *buf, int size);
+int tegra_ivc_read(struct ivc *ivc, void *buf, size_t size);
 
 /**
  * ivc_hv_ivc_can_read - Test whether data are available
@@ -175,7 +177,7 @@ int tegra_hv_ivc_dump(struct tegra_hv_ivc_cookie *ivck);
  */
 int tegra_hv_ivc_read_peek(struct tegra_hv_ivc_cookie *ivck,
 		void *buf, int off, int count);
-int tegra_ivc_read_peek(struct ivc *ivc, void *buf, int off, int count);
+int tegra_ivc_read_peek(struct ivc *ivc, void *buf, size_t off, size_t count);
 
 /**
  * ivc_hv_ivc_read_get_next_frame - Peek at the next frame to receive
@@ -214,7 +216,8 @@ int tegra_ivc_read_advance(struct ivc *ivc);
  */
 int tegra_hv_ivc_write_poke(struct tegra_hv_ivc_cookie *ivck,
 		const void *buf, int off, int count);
-int tegra_ivc_write_poke(struct ivc *ivc, const void *buf, int off, int count);
+int tegra_ivc_write_poke(struct ivc *ivc, const void *buf, size_t off,
+		size_t count);
 
 /**
  * ivc_hv_ivc_write_get_next_frame - Poke at the next frame to transmit
